+++ tests/js/session-tests.js
@@ -194,16 +194,15 @@ module.exports = {
     const partition = Utils.genPartition();
     let credentials = Realm.Credentials.anonymous();
     let app = new Realm.App(appConfig);
-
-    return new Promise((resolve, reject) => {
-      return app.logIn(credentials).then((user) => {
-        let config = getSyncConfiguration(user, partition);
-        config.sync.user = { username: "John Die" }; // this is an invalid user object
-        return Realm.open(config)
-          .then((e) => reject(JSON.stringify(e)))
-          .catch((e) => { console.log("FISK 101"); resolve() });
-      });
-    });
+    let user = app.logIn(credentials);
+    let config = getSyncConfiguration(user, partition);
+    config.sync.user = { username: "John Doe" }; // this is an invalid user object
+    try {
+      await Realm.open(config);
+      Promise.reject();
+    } catch (_) {
+      Promise.resolve();
+    }
   },
 
   testRealmOpen() {